+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Change `digits` argument to an `options` argument
+
 2.0.1 / 2014-08-10
 ==================